backend/dhobi_backend/requirements.txt
asgiref==3.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.0
